Added some content, and fixed some spelling: See extended description for more details
so i added the start of day 2, where Arshneet is in homeroom for 2 periods. he has the voice of fooling around, playing games (game over) or working (continue) if he chooses to fool around, arash tells a joke and if he laughs, it induces rage in his classmates killing him. if he doesn't then it continues to period 3.

Problem is, i need to make it so both the work option and the fool around option (assuming you don't laugh) to lead to the same code. There's no "goto" option, so i'm stumped.